+++ src/Arguments/ServiceBus/Topic/SubscriptionPeekArguments.cs
@@ -6,6 +6,11 @@ namespace AzureMcp.Arguments.ServiceBus.Subscription;
 public class SubscriptionPeekArguments : BaseTopicArguments
 {
     /// <summary>
+    /// Name of the subscription.
+    /// </summary>
+    public string? SubscriptionName { get; set; }
+
+    /// <summary>
     /// Maximum number of messages to peek from subscription.
     /// </summary>
     public int? MaxMessages { get; set; }