+++ src/Commands/ServiceBus/ServiceBusJsonContext.cs
@@ -8,8 +8,11 @@ using AzureMcp.Commands.ServiceBus.Topic;
 namespace AzureMcp.Commands.ServiceBus;
 
 [JsonSerializable(typeof(QueuePeekCommand.QueuePeekCommandResult))]
+[JsonSerializable(typeof(QueueDetailsCommand.QueueDetailsCommandResult))]
 [JsonSerializable(typeof(SubscriptionPeekCommand.SubscriptionPeekCommandResult))]
-[JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase)]
+[JsonSerializable(typeof(SubscriptionDetailsCommand.SubscriptionDetailsCommandResult))]
+[JsonSerializable(typeof(TopicDetailsCommand.TopicDetailsCommandResult))]
+[JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase, DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull)]
 internal sealed partial class ServiceBusJsonContext : JsonSerializerContext
 {
 }